[hms] relax the notification format restriction
'messageFormat' is an optional field in NotificationEvent thrift object
introduced in HIVE-10562. In order to be compatible with Hive
distributions that do not include HIVE-10652, this commit relaxes the
restriction of HMS notification message format, so that the
notification log listener can process the message even when the format
type is not provided. Removing this restriction should not cause any
regressions as long as the required fields are present in the messages.

Change-Id: I09d0e36cebe2074d975504d6bbc5677107e33fe3
Reviewed-on: http://gerrit.cloudera.org:8080/11460
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins