[test] Clean up MiniKuduCluster and BaseKuduTest
This patch cleans up the MiniKuduCluster and
BaseKuduTest API in preparation for creating a
public test API and improving per-test
MiniKuduCluster parameters.

This patch also removes the use of Guava’s
HostAndPort as part of the clean up. It replaces
it with our own HostAndPort which wraps an
unresolved InetSocketAddress. We can’t use
Guava’s HostAndPort because it is shaded and
relocated in all of our libraries. This means it shouldn’t
be a part of the “public” API. In this case that means
the test API too. Additionally HostAndPort is marked
as beta.

Change-Id: I2a55c41f48f31c4a12b41508dbb343c4419e65b1
Reviewed-on: http://gerrit.cloudera.org:8080/11474
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>