[test] Adjust Kudu binary locator logic.
This patch adjusts the binary locating logic to be less
fragile and more externally useful. It also breaks the
logic out of the TestUtils class into it’s own class.
I broke out the remaining TestUtils methods as
well and remove any that were unused.

This patch also adjusts the Maven and Gradle build
to define the `kuduBinDir` system property. Doing
this means we don’t need to detect git repos and
search the repository for binaries.  This also fixes
the issue where Maven tests would fail to locate the
binary because they were using jars installed into
the local repository.

Change-Id: I365dfb9fdeeb4fd985f370a6b8305345e0a2ac7d
Reviewed-on: http://gerrit.cloudera.org:8080/11475
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>