tserver: log a message when a scanner is not found
Right now there is a class of issues where an "upstream" service or
query engine such as Impala or Spark may not be able to find a Kudu
scanner, resulting in a failed query. Often, this is because the scanner
timed out. The query engine will receive and typically log a message
indicating that the scanner could not be found, however it's difficult
to trace this back to the original query because the client wouldn't
know the scanner id and the server would not log the event.

With this change, it will be much easier to match up query engine
failures with expiring scanners by looking at the logs on both sides
because when a request comes in for a scanner that cannot be found, the
client will get a bad status:

  Not found: Scanner 7672e46ed30d42938c54bf6e7e24946e not found (it may have expired)

And the server will log a slightly more verbose message:

  I0926 14:59:50.600463 21137 tablet_service.cc:2020] Scan: Not found: Scanner 7672e46ed30d42938c54bf6e7e24946e not found (it may have expired): call sequence id=1, remote={username='mpercy'} at 127.0.0.1:42660

Additionally, the above situation is handled in a similar way in the
case of a TabletService::ScannerKeepAlive() RPC call.

Change-Id: If4e9c69160605d4a839ac2d28def67f4d3402b90
Reviewed-on: http://gerrit.cloudera.org:8080/11516
Tested-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>