[tools] ksck checksums: Add KsckChecksummer class
This removes the remaining checksum logic out of Ksck and into a new
KsckChecksummer class, along with refactoring some parts of the logic
into separate functions. There are no functional changes.

The logic that was not refactored is directly related to KUDU-2179, and
will be addressed in a follow-up patch.

Change-Id: I2016936eaa26fd6b499783e7d5d8f404816b37fa
Reviewed-on: http://gerrit.cloudera.org:8080/11498
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>