java: improve assertion failure message in TestAsyncKuduSession.test
This test is fairly flaky (tracked in KUDU-1521) and while I don't have
time to fix the underlying issue I want to make it easier to track this
in the future. Instead of just throwing a base AssertionError, it will
now throw an error indicating that it failed because it was expecting a
PleaseThrottleException that never appeared.

Change-Id: If8b2f300f34ec70ee48d81a96f378587de9f8a04
Reviewed-on: http://gerrit.cloudera.org:8080/11510
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Mike Percy <mpercy@apache.org>