create-table-stress-test: deflake TestConcurrentCreateTableAndReloadMetadata
First, if the ASSERT_OK fired, the test would hang. No bueno.

Second, in TSAN environments the test would sometimes fail with a table
already present error, typically happening either 30s or 60s after that
table was created. I believe this is due to a client-side timeout _after_
the table was created but _before_ its RPC response was sent. We could
try to address this by bumping the client's timeout, but because
reload_metadata_thread is DoS'ing the cluster, that doesn't seem robust.
Instead, let's have the thread ease up in TSAN environments.

Without the change, the test failed 4/1000 times (--stress_cpu_threads=4).

With the change, the test failed 0/1000 times (--stress_cpu_threads=8).

Change-Id: I607a8e278e8e8f370cda512f365a97fbb81e1ae3
Reviewed-on: http://gerrit.cloudera.org:8080/11523
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins