[test] Move BaseKuduTest to a Junit Rule
This patch moves all of BaseKuduTest to a Junit Rule.
This avoids inheritance for tests and allows more
interesting test composition.

Additionally I added method level annotations
that can be used to modify the Kudu mini cluster
configs on a per test method basis. I changed
the tests in TestKuduClient to use the annotations,
and will migrate other tests in follow on patches.

Change-Id: I32c83b47a576377b924ea41dbeaf78ce2b75e4c4
Reviewed-on: http://gerrit.cloudera.org:8080/11547
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>