[master] extra tests for the placement policy
Added a few more tests for the replica placement logic.  This is to
verify how an extra replica is placed in case of a tablet with RF=5.

RF=5 test scenarios are interesting in exercising the replica placement
logic vs RF=3 scenarios because placing two replicas into one location
does not constitute a violation of placement policy in case of RF=5.
So, those RF=5 scenarios allows to verify how the placement logic works
during initial replica placement and while adding an extra replica
when it's possible to place more than one replica into same location.

This is a follow-up to ebb2852d99ed27c26e65c3569d5cb515754b2937.

Change-Id: I49748eca01743fb94846ccd611c5f1c7ddb20c04
Reviewed-on: http://gerrit.cloudera.org:8080/11562
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>