KUDU-1592 Make warnings about FBM more ominous
The documentation warned about scalability issues with file block
manager, but it was rather lightly worded and so was the warning message
of a hole punch test failure.

This commit removes the actual flag necessary to enable FBM from the
hole punching tests' warning to force users look it up in the
documentation instead of blindly applying it, adds a note about it being
unsuitable for production use to the flag's description and expands the
warning in the troubleshooting section.

Change-Id: I15adf42dd7e6376a85cdb178ad460f239a4f87a1
Reviewed-on: http://gerrit.cloudera.org:8080/11579
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Attila Bukor <abukor@apache.org>