[tests] location assignment for ExternaMiniCluster
Introduced location assignment rules for tablet servers running as
part of ExternalMiniCluster.  With this changelist, it's possible
to have tablet servers of ExternalMiniCluster to be spread among
locations in accordance with simple number-of-servers-at-location
mapping rules.

For example, the following is a self-descriptive example of such
mapping rules set to spread six tablet servers in a cluster among
three locations:

  L0:1  L1:2  L2:3

Added an integration test to verify the location assignment
rules for ExternalMiniCluster work as intended.

Change-Id: I63309804cf2fdd8a620b50abdd7133af6a033c30
Reviewed-on: http://gerrit.cloudera.org:8080/11606
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>