[java] fixed kudu-jepsen build
The code in client.clj and table.clj in the kudu-jepsen package were
importing BaseKuduTest but didn't use it.  After the recent changes,
the Clojure build cannot import BaseKuduTest anymore.  Since it was
unused anyway, I decided to stop importing the useless stuff without
digging in further.  That fixed the build error.

This is a follow-up to dc8ae79961f71b8bdc344781fc89d38d94152fc4.

Change-Id: I2d88b55a5f0b3ef4bf683d1de1bddc090d351b56
Reviewed-on: http://gerrit.cloudera.org:8080/11622
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>