client: add tablet id to scanner error messages
Some of the existing scanner log messages are confusing because at some
point the KuduScanner::Data::DebugString() method was changed to have
different output. This patch cleans up the output so that the sentences
make sense and so we always print the id of the tablet that returned the
error, if it's available.

Also fixed a couple of clang-tidy warnings.

Change-Id: I349d624c87d924df81d34c5eeb556433eac7b440
Reviewed-on: http://gerrit.cloudera.org:8080/11645
Tested-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>