[master] --location_mapping_cmd is experimental
Re-tagged --location_mapping_cmd: evolving --> experimental.

Not all pieces of the location-awareness is there yet, so let's
mark the location mapping flag as 'experimental' instead of 'evolving'.

The rationale is simple: the feature is not yet code-complete and
there might be some usability issues if using the location awareness
provisions at this point.  E.g., location-unaware rebalancer would
violate the placement policies if run against a cluster with locations
defined, etc.

Change-Id: I783f07ff070d608deacfd6dd055fede80795b3ee
Reviewed-on: http://gerrit.cloudera.org:8080/11652
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>