Remove column ids from KuduSchema::ToString
708d0f736c5a9d60ca02041bdb3b753e11478b6d added a KuduSchema::ToString
method, which calls into Schema::ToString. This is fine except
Schema::ToString includes the column ids when they are set, and that
information is implementation and ought not be shown to the user.

I think the original testing missed this because it never called
KuduSchema::ToString on a KuduSchema built from a Schema that had column
ids. A new test fixes this oversight.

Change-Id: Ib5999c485b63b6bdd4f4778b329f6768680b5193
Reviewed-on: http://gerrit.cloudera.org:8080/11668
Tested-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>