catalog_manager: adjust maximum tablet count at table creation time
The maximum tablet check introduced in commit 64983c454 was incomplete in
that it only considered the number of _tablets_ rather than the number of
_replicas_, even though replica creation itself places load on a tserver
during table creation. This patch clarifies the intended behavior and
incorporates the replication factor of the table into the check.

Note: users who override --max_create_tablets_per_ts are in for a rude
surprise as the maximum size of their tables will be effectively cut in
three. Given that the parameter isn't tagged as "stable" though, I think
communicating this change in behavior via release note is sufficient.

Change-Id: I00d91baddb1591476a7be27cba043e6354558208
Reviewed-on: http://gerrit.cloudera.org:8080/11716
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>