[tools] updated comments on 'kudu perf loadgen'
While using the 'kudu perf loadgen' tool recently, I noticed that the
descriptions for a few of its flags were not clear enough.  This changelist
fixes that and also updates the examples of usage since some of those had
the masters' RPC end-point argument misplaced.

This changelist does not contain any functional changes.

Change-Id: If171957db9d9be8d14b5031a64f4edbfa5470be0
Reviewed-on: http://gerrit.cloudera.org:8080/11724
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>