[tools] Add locate_row support for unsupported key column types
BOOL, FLOAT, and DOUBLE are not supported as types for key columns. But
there's actually no reason why the tool can't pre-emptively support them
in case they are ever permitted as key column types.

There's no additional tests because it's not possible to create a table
with these types in the primary key.

Change-Id: Id1ba6de90feef4c2c565d9033ff442b723a51d7b
Reviewed-on: http://gerrit.cloudera.org:8080/11729
Tested-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>