[build] Centralize project version to version.txt
This patch changes the Java and Python builds
to use the version.txt as their source of the project
version. This simplifies branching and ensures
our versions always match.

I use symbolic links in the python project to ensure
that the version and licence file are copied from the
parent directory when generating a source
distribution.

This patch also corrects the versioning of snapshot
builds to ensure they include .dev0. This was being
done in the kudu/version.py file, but not in the
distribution version set in setup(…).

Change-Id: Ib654be8657ae3dc2f193056484b5d0743aa1ce95
Reviewed-on: http://gerrit.cloudera.org:8080/11735
Tested-by: Kudu Jenkins
Reviewed-by: <andy@phdata.io>
Reviewed-by: Adar Dembo <adar@cloudera.com>