De-flake ASAN TestSimultaneousLeaderTransferAndAbruptStepdown
This test is flaky in ASAN when run with the entire kudu-admin-test,
though not when run by itself. I'm not sure why that is, but to try to
improve things, I'm reducing the frequency of the simultaneous abrupt
stepdown and graceful transfer in ASAN mode so there's more time for
the tablet to make progress.

I ran the new test as part of the entire kudu-admin-test 1000 times in
ASAN and saw no failures. I ran kudu-admin-test 1000 times in ASAN with
8 stress threads and saw 4 failures, none of which were
TestSimultaneousLeaderTransferAndAbruptStepdown.

Change-Id: Ic3237f3fde48daeb2f307dfb78d457217df2beed
Reviewed-on: http://gerrit.cloudera.org:8080/11737
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>