KUDU-2542: add initial authorization token impl
This patch adds an authorization token that echoes the authentication
token implementation. These tokens contain privileges that will be used
authorize specific tablet server requests.

By in large, tablet server requests are scoped to a single table, and as
such, so are authz tokens. In cases where this is not true (e.g.
ListTablets), a reasonable assumption is that the call is being made via
tooling, and coarse-grained access control should be used instead of
fine-grained. If this ends up being less the case in the future, we can
always amend the authz token to support multiple tables.

The tokens leverage the same token signer as the authentication tokens,
though with the token validity interval configured via the new flag
--authz_token_validity_seconds.

Change-Id: Id28747ec38675abdf50dce1e7c176d29213e370f
Reviewed-on: http://gerrit.cloudera.org:8080/11750
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>